Convert most containers to use common compilation container (#6)
Converting the existing containers to use a common compilation container with pre-built libraries and standard compilers.

This effort also re-arranged the common directory structure of the different software and data containers. Now rather than a large number of top-level directories containing software, data, and scripts, the data that is not part of the operating system can now be found in three top-level directories:

/comsoftware - contains software packages that are built in each container (e.g. /comsoftware/wrf, /comsoftware/met), as well as libraries that are specific to community software (e.g. netCDF, BUFRLIB) which are built in the common compilation container under /comsoftware/libs
/data - contains input data for different packages. Data containers will be mounted here.
/home - contains scripts and working directories for each piece of software. Working directories (e.g. wpsprd, wrfprd, gsiprd, etc.) will be mounted here.
I also standardized the container permissions into a user known as 'comuser', rather than always running as root. This is considered best practice for security purposes (https://opensource.com/article/18/3/just-say-no-root-containers). This user has write permissions under /home, /comsoftware, and /data.

Runtime instructions needed to be changed for this new directory structure, and can be found in the updated README files.