Change key servers to point to an active server, hkps://keys.openpgp.org (#50)
The keyserver pool.sks-keyservers.net, which I had been using to verify the download of the "gosu" tool in the base images, is no longer active. So when trying to build the base image, this step fails. This is summarized in this issue:
 - https://github.com/docker-library/mysql/issues/263

This is mentioned in several places across the internet, including  https://www.reddit.com/r/crypto/comments/o7oh4w/skskeyserversnet_pool_dns_records_disabled/ . Even though this keyserver has been deprecated for a while, there's a lot of old stale documentation pointing to it, including that which I originally based by entrypoint script on.

The fix is simply to point to an active, more reliable keyserver,  hkps://keys.openpgp.org. This will fix the build issue, although it apparently remains to be seen what the future holds for key signing in this way. 

Still unsure if this is related to #46 or not....but hopefully when this is fixed we will see more reliable behavior?